feat(config): add skip_afk_prompt_injection option to suppress AFK mode reminders refactor(yolo): remove YoloModeInjectionProvider and related logic test: update tests for skip_afk_prompt_injection and remove yolo injection tests
