+++ docs/en/configuration/config-files.md
@@ -27,6 +27,7 @@ The configuration file contains the following top-level configuration items:
 | `default_model` | `string` | Default model name, must be a model defined in `models` |
 | `default_thinking` | `boolean` | Whether to enable thinking mode by default (defaults to `false`) |
 | `default_yolo` | `boolean` | Whether to enable YOLO (auto-approve) mode by default (defaults to `false`) |
+| `skip_afk_prompt_injection` | `boolean` | Whether to suppress the AFK mode system reminder (defaults to `false`) |
 | `default_plan_mode` | `boolean` | Whether to start new sessions in plan mode by default (defaults to `false`); resumed sessions preserve their existing state |
 | `default_editor` | `string` | Default external editor command (e.g. `"vim"`, `"code --wait"`), auto-detects when empty |
 | `theme` | `string` | Terminal color theme, either `"dark"` or `"light"` (defaults to `"dark"`) |
@@ -45,6 +46,7 @@ The configuration file contains the following top-level configuration items:
 default_model = "kimi-for-coding"
 default_thinking = false
 default_yolo = false
+skip_afk_prompt_injection = false
 default_plan_mode = false
 default_editor = ""
 theme = "dark"
