+++ docs/en/release-notes/changelog.md
@@ -4,15 +4,15 @@ This page documents the changes in each Kimi Code CLI release.
 
 ## Unreleased
 
-- Core: Fix `--yolo` mode unintentionally preventing the model from calling `AskUserQuestion` — yolo used to inject a system reminder telling the model it was in "non-interactive mode" and must not ask, and the ask-user tool auto-dismissed in yolo. Both were wrong: yolo only means "auto-approve tool calls", not "the user is gone". Yolo now keeps the user fully reachable, with a gentler prompt that tells the model it can still ask clarifying questions at meaningful decision points (just to use it sparingly)
+- Core: Fix `--yolo` mode unintentionally preventing the model from calling `AskUserQuestion` — yolo used to inject a system reminder telling the model it was in "non-interactive mode" and must not ask, and the ask-user tool auto-dismissed in yolo. Both were wrong: yolo only means "auto-approve tool calls", not "the user is gone". Yolo is now purely a runtime approval policy and does not inject model guidance; the user remains reachable through `AskUserQuestion`
 - CLI: Split "auto-approve" and "non-interactive" into two orthogonal modes — `--yolo` (auto-approve tool calls while the user is still at the terminal) and `--afk` / `/afk` (away-from-keyboard: `AskUserQuestion` is auto-dismissed and tool calls are auto-approved). `--print` now implicitly enables `--afk` instead of `--yolo` (more accurate: print has no user at the terminal). The status bar shows `yolo` and `afk` independently, and `/yolo` and `/afk` toggle their own flag without disturbing the other
+- Config: Rename `skip_yolo_prompt_injection` to `skip_afk_prompt_injection` now that yolo no longer injects model guidance. The old config key is ignored if present
 - Shell: Fix `/yolo` toggling producing misleading UI messages when afk is also active — `/yolo` used to read the combined auto-approve state, so pressing it under afk would claim approval was now required even though afk kept auto-approving everything. `/yolo` now reads and writes only the yolo flag, leaving afk alone
 - Web: Fix AI title generation overwriting a manually-set title when the LLM call finishes after the user has already renamed the session — the final write now reloads state and yields to a `title_generated` flag set by another request
 - Web: Surface session rename, archive, unarchive, and title generation failures as toast notifications instead of only logging to the console
 - Web: Keep tool media previews visible when tool details are collapsed — images and videos returned by tools now render below the tool card instead of inside the collapsible detail area, so preview thumbnails remain accessible after collapsing a tool
 - Kosong: Fix stale API key after OAuth token refresh in Kimi provider — `on_retryable_error` now reads the current `api_key` from the live client instead of the cached `_api_key`, so that OAuth token refreshes applied via `client.api_key` are preserved when the client is rebuilt after a retryable error
 - Core: Approval requests no longer auto-timeout after 5 minutes, which previously surfaced as `Rejected by user`; active foreground and subagent approvals now wait indefinitely for user response
-- Core: Fix yolo mode reminder being lost after context compaction — the non-interactive-mode guidance ("don't call AskUserQuestion, plan-mode toggles are auto-approved") is now re-injected on the first LLM step after each compaction while yolo remains active, instead of being silently dropped when the original reminder is folded into the compaction summary
 - Shell: Fix `/usage` remaining quota rendering — the progress bar, warning colors, and `% left` label now all use the remaining quota ratio consistently, so high remaining quota shows as green/full and near-exhausted quota shows as yellow or red
 - Shell: Show active background agent task count in the prompt status bar — the existing `⚙ bash: N` badge only counted background Shell tasks and filtered out background Agent subagents, so when many subagents were running the prompt looked idle and users could not tell work was in progress; the toolbar now renders `⚙ bash: N` and `⚙ agent: N` as two independent badges (each hidden when its count is 0) and drops the agent badge first when the terminal is too narrow to fit both
 - Auth: Fix managed model list refresh silently failing for OAuth users with expired tokens — the background `/models` sync now detects 401 responses, forces an OAuth token refresh, and retries with the refreshed token; if the refresh fails or the refreshed token is still rejected, it falls back to the originally configured static API key instead of skipping the provider
