+++ src/kimi_cli/config.py
@@ -199,12 +199,11 @@ class Config(BaseModel):
     default_model: str = Field(default="", description="Default model to use")
     default_thinking: bool = Field(default=False, description="Default thinking mode")
     default_yolo: bool = Field(default=False, description="Default yolo (auto-approve) mode")
-    skip_yolo_prompt_injection: bool = Field(
+    skip_afk_prompt_injection: bool = Field(
         default=False,
         description=(
-            "If true, suppress the system reminders that are normally injected when yolo "
-            "or afk mode is active. Useful when building custom applications on top of "
-            "KimiSoul that do not need the non-interactive mode hint."
+            "If true, suppress the afk-mode system reminder. "
+            "Yolo mode does not inject a system reminder."
         ),
     )
     default_plan_mode: bool = Field(default=False, description="Default plan mode for new sessions")
