+++ tests/core/test_skip_afk_prompt_injection.py
@@ -1,7 +1,7 @@
-"""Tests for the `skip_yolo_prompt_injection` config gate.
+"""Tests for the `skip_afk_prompt_injection` config gate.
 
-When enabled, both `YoloModeInjectionProvider` and `AfkModeInjectionProvider`
-should be left out of the injection pipeline. Plan mode injection is unaffected.
+Yolo no longer has a dynamic prompt. This field gates only the afk prompt
+provider. Plan mode injection is unaffected.
 """
 
 from __future__ import annotations
@@ -15,7 +15,6 @@ from kimi_cli.soul.agent import Agent, Runtime
 from kimi_cli.soul.context import Context
 from kimi_cli.soul.dynamic_injections.afk_mode import AfkModeInjectionProvider
 from kimi_cli.soul.dynamic_injections.plan_mode import PlanModeInjectionProvider
-from kimi_cli.soul.dynamic_injections.yolo_mode import YoloModeInjectionProvider
 from kimi_cli.soul.kimisoul import KimiSoul
 
 
@@ -35,19 +34,18 @@ def _provider_types(soul: KimiSoul) -> set[type]:
 
 
 @pytest.mark.parametrize("skip", [False, True])
-def test_skip_yolo_prompt_injection_gates_both_providers(
+def test_skip_afk_prompt_injection_gates_afk_provider(
     runtime: Runtime, tmp_path: Path, skip: bool
 ) -> None:
-    runtime.config.skip_yolo_prompt_injection = skip
+    runtime.config.skip_afk_prompt_injection = skip
     soul = _make_soul(runtime, tmp_path)
     types_ = _provider_types(soul)
 
     # Plan is always present and never gated by this flag.
     assert PlanModeInjectionProvider in types_
+    assert not any(provider.__name__.lower().startswith("yolo") for provider in types_)
 
     if skip:
-        assert YoloModeInjectionProvider not in types_
         assert AfkModeInjectionProvider not in types_
     else:
-        assert YoloModeInjectionProvider in types_
         assert AfkModeInjectionProvider in types_
